# reduceNonRepPrim: postpone call to shouldReduce